requirements.txt
mcp>=1.0.0
httpx>=0.25.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0